Create useMovieQuery hook to handle API fetch
UseMovieQuery provides search function and load more function, which take the search query and request list of movies from TMDB. Results are paginated to reduce the amount of movies loaded into the app at a time. Also provides error, loading and hasMoreMovies states to be used to alter UI based on status of data fetch.