Add useMovieDetails hook to fetch data for in-depth movie view
useMovieDetails will be used to populate the view that will appear when user clicks on a movie from the list. Multiple requests from movie details, videos, and credits to obtain general details as well as top cast members and the trailer. Moved api key and base url out to contants file for reusability.